requirements.txt
asn1crypto==0.24.0
cached-property==1.4.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
crypto==1.4.1
cryptography==2.2.2
idna==2.7
Naked==0.1.31
pycparser==2.18
pycryptodomex==3.6.3
PyKCS11==1.5.2
pykeytool==0.1
pyOpenSSL==18.0.0
PyYAML==3.12
requests==2.19.1
shellescape==3.4.1
six==1.11.0
urllib3==1.23
